remove server logs for web
